Extract common FreshBooks code to lib/freshbooks
